requirements.txt
Click==7.0
Flask==1.0.3
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
pkg-resources==0.0.0
Werkzeug==0.15.4
